🐛 fix: 파일 삭제 시 경로 문제 해결 및 파일 없을 경우 예외 발생 제거

+++ server/src/user/service/user.service.ts
@@ -309,7 +309,9 @@ export class UserService {
       updateData.profileImage !== undefined &&
       user.profileImage !== updateData.profileImage
     ) {
-      await this.fileService.deleteByPath(user.profileImage);
+      if (user.profileImage) {
+        await this.fileService.deleteByPath(user.profileImage);
+      }
       user.profileImage = updateData.profileImage;
     }
     if (updateData.introduction !== undefined) {
